Create Prime
To check if a number is prime, the program tests if the number has any divisors other than 1 and itself. It does this by iterating through all numbers from 2 to the square root of the given number. If any number divides evenly into the input number, it is not prime. Otherwise, it is prime.