Create Matrix
Matrix multiplication involves taking two matrices and producing a third matrix. Each element of the resulting matrix is calculated as the dot product of corresponding rows and columns from the two matrices. The number of columns in the first matrix must be equal to the number of rows in the second matrix for multiplication to be possible.